+++ src/main/java/de/presti/ree6/main/Main.java
@@ -30,14 +30,16 @@ import de.presti.ree6.sql.entities.ScheduledMessage;
 import de.presti.ree6.sql.entities.Setting;
 import de.presti.ree6.sql.entities.TwitchIntegration;
 import de.presti.ree6.sql.entities.stats.ChannelStats;
-import de.presti.ree6.sql.entities.stats.Statistics;
 import de.presti.ree6.sql.util.SQLConfig;
 import de.presti.ree6.sql.util.SettingsManager;
 import de.presti.ree6.utils.apis.ChatGPTAPI;
 import de.presti.ree6.utils.apis.Notifier;
 import de.presti.ree6.utils.apis.SpotifyAPIHandler;
-import de.presti.ree6.utils.data.*;
+import de.presti.ree6.utils.config.Config;
+import de.presti.ree6.utils.data.ArrayUtil;
 import de.presti.ree6.utils.external.RequestUtility;
+import de.presti.ree6.utils.oauth.CustomOAuth2Credential;
+import de.presti.ree6.utils.oauth.CustomOAuth2Util;
 import de.presti.ree6.utils.others.ThreadUtil;
 import io.sentry.Sentry;
 import lavalink.client.io.jda.JdaLavalink;
